# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Garagify
  version: 1.0.0
  description: API
tags:
  - name: Authorization
    description: Operations related to user authentication and authorization
  - name: Lend
    description: Lend out parking spots
  - name: Borrow
    description: Borrow a parking spot
  - name: Parking
    description: Parking and spot management
  - name: LendOfferController
    description: Legacy controller for lend offers

paths:
  /api/v1/auth/access:
    post:
      tags:
        - Authorization
      summary: Authenticate using a Google token
      operationId: authenticateWithGoogle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenVerificationForm'
      responses:
        '200':
          description: JWT tokens returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenDto'
  /api/v1/auth/refresh:
    post:
      tags:
        - Authorization
      summary: Refresh JWT token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenVerificationForm'
      responses:
        '200':
          description: New JWT tokens returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenDto'
  /api/v1/parkings/{parking_id}/lend/{spot_id}:
    post:
      summary: Borrow a specific parking spot
      tags:
        - Lend
      operationId: createLendForSpot
      parameters:
        - name: parking_id
          in: path
          required: true
          schema:
            type: integer
        - name: spot_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeRangeRequest'
      responses:
        '201':
          description: Borrow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendOfferDTO'
  /api/v1/lend/mine:
    get:
      operationId: getMyLends
      summary: Get all borrows for the current user (paged)
      tags:
        - Lend
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number (0-based)
        - in: query
          name: size
          schema:
            type: integer
          description: Page size
        - in: query
          name: sort
          schema:
            type: string
          description: Sort order (e.g., borrowTime,asc)
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendOfferListDTO'
  /api/v1/lend/{id}:
    delete:
      tags:
        - Lend
      summary: Delete a borrow by its ID
      description: Deletes the lend resource with the given UUID.
      operationId: deleteParkingLend
      parameters:
        - name: id
          in: path
          description: UUID of the lend to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Borrow deleted successfully, no content returned
  /api/v1/parkings/{parking_id}/borrow/{spot_id}:
    post:
      summary: Borrow a specific parking spot
      tags:
        - Borrow
      operationId: createBorrowForSpot
      parameters:
        - name: parking_id
          in: path
          required: true
          schema:
            type: integer
        - name: spot_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeRangeRequest'
      responses:
        '201':
          description: Borrow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowDTO'
  /api/v1/borrow/mine:
    get:
      operationId: getMyBorrows
      summary: Get all borrows for the current user (paged)
      tags:
        - Borrow
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number (0-based)
        - in: query
          name: size
          schema:
            type: integer
          description: Page size
        - in: query
          name: sort
          schema:
            type: string
          description: Sort order (e.g., borrowTime,asc)
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowListDTO'
  /api/v1/borrow/{id}:
    delete:
      tags:
        - Borrow
      summary: Delete a borrow by its ID
      description: Deletes the borrow resource with the given UUID.
      operationId: deleteParkingBorrow
      parameters:
        - name: id
          in: path
          description: UUID of the borrow to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Borrow deleted successfully, no content returned
  /api/v1/parkings/{parking_id}:
    get:
      operationId: getParking
      summary: Get parking by id
      tags:
        - Parking
      parameters:
        - in: path
          name: parking_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingDTO'
        404:
          description: Not found
  /api/v1/parkings/{parking_id}/lend/available-timeranges:
    get:
      tags:
        - Parking
      summary: Get available time ranges for lending a parking spot
      description: Returns an array of available time ranges during which a parking spot can be lent up to the given date.
      operationId: getAvailableLendTimeRanges
      parameters:
        - in: path
          name: parking_id
          required: true
          schema:
            type: integer
        - in: query
          name: untilWhen
          required: true
          schema:
            type: string
            format: date-time
            example: '2025-06-01T23:59:00+02:00'
          description: Date (inclusive) until when the parking spot can be lent.
      responses:
        '200':
          description: Array of available time ranges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeRangeDto'

  /api/v1/parkings/{parking_id}/lend/available-spots:
    get:
      tags:
        - Parking
      summary: Get available parking spot UUIDs for lending in a given time range
      description: Returns a list of spot UUIDs that are available to lend for the entire given time range.
      operationId: getAvailableLendSpots
      parameters:
        - in: path
          name: parking_id
          required: true
          schema:
            type: integer
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
            example: '2025-06-01T09:00:00+02:00'
          description: Start of the time range (with offset)
        - in: query
          name: until
          required: true
          schema:
            type: string
            format: date-time
            example: '2025-06-01T18:00:00+02:00'
          description: End of the time range (with offset)
      responses:
        '200':
          description: List of available spot UUIDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
  /api/v1/parkings/{parking_id}/borrow/available-timeranges:
    get:
      tags:
        - Parking
      summary: Get available time ranges for borrowing a parking spot
      description: Returns an array of available time ranges during which a parking spot can be borrowed up to the given date.
      operationId: getAvailableBorrowTimeRanges
      parameters:
        - in: path
          name: parking_id
          required: true
          schema:
            type: integer
        - in: query
          name: untilWhen
          required: true
          schema:
            type: string
            format: date-time
            example: '2025-06-01'
          description: Date (inclusive) until when the parking spot can be borrowed.
      responses:
        '200':
          description: Array of available time ranges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeRangeDto'
  /api/v1/parkings/{parking_id}/borrow/available-spots:
    get:
      tags:
        - Parking
      summary: Get available parking spot UUIDs for a given time range
      description: Returns a list of spot UUIDs that are available for the entire given time range.
      operationId: getAvailableBorrowSpots
      parameters:
        - in: path
          name: parking_id
          required: true
          schema:
            type: integer
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
            example: '2025-06-01T09:00:00+02:00'
          description: Start of the time range (with offset)
        - in: query
          name: until
          required: true
          schema:
            type: string
            format: date-time
            example: '2025-06-01T18:00:00+02:00'
          description: End of the time range (with offset)
      responses:
        '200':
          description: List of available spot UUIDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
  /api/v1/lend-offer:
    post:
      summary: Create a lend offer
      tags:
        - LendOfferController
      description: This is a description
      operationId: createLendOffer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LendOfferPostForm'
      responses:
        200:
          description: This means its ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendOfferDTO'
    get:
      summary: Lend offers by filters
      tags:
        - LendOfferController
      description: This is a description
      operationId: getAllLendOffers
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: Page number (zero-based)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: Number of items per page
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Start date for the date range filter (inclusive)
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: End date for the date range filter (inclusive)
        - name: owner_id
          in: query
          required: false
          schema:
            type: string
          description: The id of the owner to filter by
      responses:
        200:
          description: This means its ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendOfferListDTO'
  /api/v1/lend-offer/{lend_offer_id}:
    get:
      summary: Lend offer by id
      tags:
        - LendOfferController
      description: This is a description
      operationId: getLendOfferById
      parameters:
        - name: lend_offer_id
          in: path
          required: true
          description: The id of the lend offer
          schema:
            type: string
      responses:
        200:
          description: This means its ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendOfferDTO'
    put:
      summary: Update lend offer by id
      tags:
        - LendOfferController
      description: Update an existing lend offer
      operationId: updateLendOfferById
      parameters:
        - name: lend_offer_id
          in: path
          required: true
          description: The id of the lend offer
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LendOfferPutForm'
      responses:
        200:
          description: Successfully updated lend offer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendOfferDTO'
        400:
          description: Invalid input
        404:
          description: Lend offer not found
    delete:
      summary: Lend offer by id
      tags:
        - LendOfferController
      description: This is a description
      operationId: deleteLendOfferById
      parameters:
        - name: lend_offer_id
          in: path
          required: true
          description: The id of the lend offer
          schema:
            type: string
      responses:
        200:
          description: This means its ok
components:
  schemas:
    TimeRangeRequest:
      type: object
      properties:
        fromWhen:
          type: string
          format: date-time
          example: "2025-06-01T08:00:00+02:00"
        untilWhen:
          type: string
          format: date-time
          example: "2025-06-01T10:00:00+02:00"
      required:
        - fromWhen
        - untilWhen
    TokenVerificationForm:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    TokenDto:
      type: object
      properties:
        access:
          type: string
        refresh:
          type: string
      required:
        - access
        - refresh
    BorrowDTO:
      description: This is the borrow model
      type: object
      properties:
        id:
          description: Id of borrow
          type: string
        parking_id:
          description: Id of the parking
          type: integer
        spot_id:
          description: Id of parking spot
          type: string
        start_date:
          description: Start date of borrow
          type: string
          format: date-time
        end_date:
          description: End date of borrow
          type: string
          format: date-time
    BorrowForm:
      type: object
      properties:
        spotId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required:
        - spotId
        - startDate
        - endDate
    LendOfferDTO:
      description: This is the lend offer model
      type: object
      properties:
        id:
          description: Id of lend offer
          type: string
        start_date:
          description: Start date of lend offer
          type: string
          format: date-time
        end_date:
          description: End date of lend offer
          type: string
          format: date-time
        parking_id:
          description: Id of the parking
          type: integer
        spot_id:
          description: Id of parking spot
          type: string
    LendOfferPutForm:
      description: This is the lend offer put model
      type: object
      properties:
        start_date:
          description: Start date of lend offer
          type: string
          format: date-time
        end_date:
          description: End date of lend offer
          type: string
          format: date-time
        spot_id:
          description: Id of parking spot
          type: string
        owner_id:
          description: Id of user
          type: string
      required:
        - start_date
        - spot_id
        - end_date
    LendOfferPostForm:
      description: This is the lend offer model
      type: object
      properties:
        start_date:
          description: Start date of lend offer
          type: string
          format: date-time
        end_date:
          description: End date of lend offer
          type: string
          format: date-time
        spot_id:
          description: Id of parking spot
          type: string
      required:
        - start_date
        - spot_id
        - end_date
    ParkingDTO:
      description: This is the parking model (read-only)
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Unique parking identifier
        organisation_id:
          type: integer
          format: int32
          description: Id of the organisation (foreign key)
        name:
          type: string
          description: The name of the parking
        ui_object:
          type: object
          description: UI JSON data (arbitrary structure)
      required:
        - id
        - organisation_id
        - name
        - ui_object
    ParkingSpotDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Unique spot identifier
        parking_id:
          type: integer
          format: int32
          description: Id of the parent parking
        spot_uuid:
          type: string
          format: uuid
          description: UUID of the parking spot
        owner_id:
          type: string
          format: uuid
          description: Owner's user id (nullable if no owner)
        status:
          type: string
          description: Status of the parking spot (optional)
      required:
        - id
        - parking_id
        - spot_uuid
    TimeRangeDto:
      type: object
      properties:
        start:
          type: string
          format: date-time
          example: '2025-06-01T08:00:00'
        end:
          type: string
          format: date-time
          example: '2025-06-01T12:00:00'
      required:
        - start
        - end

    LendOfferListDTO:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/LendOfferDTO'
    BorrowListDTO:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/BorrowDTO'
    PagedResponse:
      description: Generic paginated response wrapper
      type: object
      properties:
        content:
          type: array
          items: { }
          description: List of items for current page
        totalElements:
          type: integer
          format: int64
          description: Total number of elements across all pages
          example: 3
        totalPages:
          type: integer
          format: int32
          description: Total number of pages
          example: 1
        page:
          type: integer
          format: int32
          description: Current page number (0-based)
          example: 0
        size:
          type: integer
          format: int32
          description: Number of elements per page
          example: 10
      required:
        - content
        - totalElements
        - totalPages
        - page
        - size